+++ index.js
@@ -5,7 +5,7 @@
 import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
-import LoginForm from './src/components/LoginForm';
+
 
 import temp from './temp';
 AppRegistry.registerComponent(appName, () => App);
