+++ src/navigation/Navigation.js
@@ -1,11 +1,9 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import LoginScreen from '@src/screens/LoginScreen';
-import HomeScreen from '@src/screens/HomeScreen'; // Add this import
 import { ROUTE } from './constant';
 import AuthStackNavigator from './AuthStackNavigator';
 import TabStackNavigator from './TabStackNavigator';
-import Splash from '@src/navigation/splash';
+import Splash from './splash';
 
 const Stack = createNativeStackNavigator();
