+++ src/screens/EventHoliday/Components/Calendar/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+import colors from "@src/theme/colors";
+import { scale } from "react-native-size-matters";
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: colors.WHITE,
+      borderBottomLeftRadius: scale(30), // Border radius at top-left
+      borderBottomRightRadius: scale(30),
+      overflow: 'hidden', // Ensure border radius is respected
+      
+      
+    },
+  });
+  export default styles;
